# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Finanças Pessoais API",
        "description": "API de controle financeiro pessoal com autenticação por token Sanctum e verificação de dois fatores (2FA).",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "Servidor Principal"
        }
    ],
    "paths": {
...
        "/api/transactions": {
            "get": {
                "tags": [
                    "Transações"
                ],
                "summary": "Listar todas as transações",
                "description": "Retorna a lista de transações (despesas e receitas) do usuário autenticado, ordenada por data em ordem decrescente.",
                "operationId": "9b50e29e364179ebde24ac5fc59cd442",
                "parameters": [
                    {
                        "name": "type",
                        "in": "query",
                        "description": "Filtra por tipo de transação ('out' para despesa, 'income' para receita)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "out",
                                "income"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de transações.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "string",
                                                        "format": "uuid",
                                                        "example": "e440e061-0b86-4fb4-8975-d91d17983637"
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Combustível"
                                                    },
                                                    "value": {
                                                        "type": "string",
                                                        "example": "150.00"
                                                    },
                                                    "type": {
                                                        "type": "string",
                                                        "example": "out"
                                                    },
                                                    "expense_date": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2026-06-09T00:00:00.000000Z"
                                                    },
                                                    "category": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "string",
                                                                "format": "uuid",
                                                                "example": "a9ef52eb-5460-449e-b9ef-dcd41bb0b793"
                                                            },
                                                            "name": {
                                                                "type": "string",
                                                                "example": "Transporte"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Transações"
                ],
                "summary": "Criar uma nova transação",
                "description": "Cadastra uma transação (despesa ou receita) para o usuário autenticado.",
                "operationId": "54eaa3952808f639d46f969f4c8b98c5",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "type",
                                    "name",
                                    "value",
                                    "expense_date",
                                    "category_id"
                                ],
                                "properties": {
                                    "type": {
                                        "type": "string",
                                        "example": "out",
                                        "enum": [
                                            "out",
                                            "income"
                                        ]
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "Almoço de Negócios"
                                    },
                                    "value": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 45.5
                                    },
                                    "expense_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-06-09"
                                    },
                                    "category_id": {
                                        "type": "string",
                                        "format": "uuid",
                                        "example": "a9ef52eb-5460-449e-b9ef-dcd41bb0b793"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Restaurante Sabor"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Transação criada com sucesso.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Transação criada com sucesso."
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "string",
                                                    "format": "uuid",
                                                    "example": "f550e061-0b86-4fb4-8975-d91d17983637"
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Almoço de Negócios"
                                                },
                                                "value": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 45.5
                                                },
                                                "type": {
                                                    "type": "string",
                                                    "example": "out"
                                                },
                                                "expense_date": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2026-06-09T00:00:00.000000Z"
                                                },
                                                "category_id": {
                                                    "type": "string",
                                                    "format": "uuid",
                                                    "example": "a9ef52eb-5460-449e-b9ef-dcd41bb0b793"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado."
                    },
                    "422": {
                        "description": "Erro de validação nos campos informados."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Insira o token recebido no fluxo de verificação 2FA (formato: Bearer <token>)",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}